Add filtering logic and refactor code
- Implemented filtering logic using Riverpod to dynamically update the restaurant list based on the search query.
- Refactored the main.dart file by moving the RestaurantListScreen to a separate file for better code organization.
- Tested the filtering functionality to ensure it accurately reflects changes in the search query.
